updated grenade schema (is_favorite field). fixed all related modules
