+++ tg-frontend/src/entities/grenade/api/__test-handlers.ts
@@ -1,6 +1,6 @@
 import { delay, http, HttpResponse } from "msw"
+import { grenadeDTOmock, grenadesDTOmock } from "../model/__mocks"
 import { api } from "./client"
-import { mockGrenade, mockGrenades } from "./__mocks"
 import { BASE_BACKEND_URL } from "@shared/config/constants"
 
 export const testGrenadeServer = ({
@@ -13,12 +13,12 @@ export const testGrenadeServer = ({
     http.get(`${BASE_BACKEND_URL}/${api.baseApiUrl}/${grenadeId}`, async () => {
         await delay(delayInMs)
 
-        return HttpResponse.json(mockGrenade)
+        return HttpResponse.json(grenadeDTOmock)
     })
 
 export const testGrenadesServer = (delayInMs: number) =>
     http.get(`${BASE_BACKEND_URL}/${api.baseApiUrl}`, async () => {
         await delay(delayInMs)
 
-        return HttpResponse.json(mockGrenades)
+        return HttpResponse.json(grenadesDTOmock)
     })
