--- tg-frontend/src/entities/grenade/ui/grenades-list/__mocks.ts
@@ -1,89 +1,0 @@
-import { GrenadeModel } from "../../model/domain"
-
-export const grenadesMocks: GrenadeModel[] = [
-    {
-        grenadeId: 1,
-        mapId: 1,
-        typeId: 1,
-        grenadeClass: {
-            name: "Smoke",
-            description: "Blocks vision temporarily",
-            price: 300,
-        },
-        propertyList: [
-            { key: "bounce", value: "medium" },
-            { key: "duration", value: "18s" },
-        ],
-        linkToVideo: "https://example.com/video1",
-        userId: 101,
-        createdAt: "2025-04-01T10:00:00Z",
-        title: "Mid Control Smoke",
-        description: "Useful for taking mid safely.",
-        isApproved: true,
-        views: 230,
-        previewImageLink: "https://example.com/image1.jpg",
-    },
-    {
-        grenadeId: 2,
-        mapId: 2,
-        typeId: 2,
-        grenadeClass: {
-            name: "Flash",
-            description: "Blinds enemies temporarily",
-            price: 200,
-        },
-        propertyList: [
-            { key: "radius", value: "15m" },
-            { key: "duration", value: "2.5s" },
-        ],
-        linkToVideo: "https://example.com/video2",
-        userId: 102,
-        createdAt: "2025-03-22T14:35:00Z",
-        title: "A Site Entry Flash",
-        description: "Perfect flash for rushing A.",
-        isApproved: true,
-        views: 500,
-        previewImageLink: "https://example.com/image2.jpg",
-    },
-    {
-        grenadeId: 3,
-        mapId: 1,
-        typeId: 3,
-        grenadeClass: {
-            name: "Molotov",
-            description: "Deals damage over time",
-            price: 400,
-        },
-        propertyList: [
-            { key: "burnTime", value: "7s" },
-            { key: "radius", value: "4m" },
-        ],
-        linkToVideo: null,
-        userId: 103,
-        createdAt: "2025-02-11T08:20:00Z",
-        title: "Post-Plant Molotov",
-        description: null,
-        isApproved: false,
-        views: 150,
-        previewImageLink: null,
-    },
-    {
-        grenadeId: 4,
-        mapId: 3,
-        typeId: 1,
-        grenadeClass: {
-            name: "Smoke",
-            description: "Obscures vision",
-            price: 300,
-        },
-        propertyList: [{ key: "duration", value: "20s" }],
-        linkToVideo: "https://example.com/video4",
-        userId: 104,
-        createdAt: "2025-01-01T12:00:00Z",
-        title: "B Site Cross Smoke",
-        description: "Classic lineup for B cross.",
-        isApproved: true,
-        views: 800,
-        previewImageLink: "https://example.com/image4.jpg",
-    },
-]
