--- tg-frontend/src/entities/map/api/__mocks.ts
@@ -1,179 +1,0 @@
-import { z } from "zod"
-import { mapDTOschema, mapPageDTOschema } from "../model"
-
-export const mockMaps: z.infer<ReturnType<typeof mapDTOschema.array>> = [
-    {
-        map_id: 1,
-        name: "Dust II",
-        link: "https://example.com/dust2",
-        image_link: "https://example.com/images/dust2.jpg",
-    },
-    {
-        map_id: 2,
-        name: "Inferno",
-        link: "https://example.com/inferno",
-        image_link: "https://example.com/images/inferno.jpg",
-    },
-    {
-        map_id: 3,
-        name: "Mirage",
-        link: null, // Отсутствует ссылка
-        image_link: "https://example.com/images/mirage.jpg",
-    },
-    {
-        map_id: 4,
-        name: "Nuke",
-        link: "https://example.com/nuke",
-        image_link: null, // Отсутствует изображение
-    },
-    {
-        map_id: 5,
-        name: "Overpass",
-        link: "https://example.com/overpass",
-        image_link: "https://example.com/images/overpass.jpg",
-    },
-]
-
-export const mockMapPage: z.infer<typeof mapPageDTOschema> = {
-    image_link:
-        "https://www.google.com/imgres?q=grenade%20image&imgurl=https%3A%2F%2Fupload.wikimedia.org%2Fwikipedia%2Fcommons%2F2%2F20%2FMkII_07.JPG&imgrefurl=https%3A%2F%2Fen.wikipedia.org%2Fwiki%2FMk_2_grenade&docid=rOjuf5ciTh6aqM&tbnid=hOkxScoNe7UFGM&vet=12ahUKEwja-Ka-zLSMAxXiHBAIHb1RGLAQM3oECBcQAA..i&w=1200&h=1638&hcb=2&itg=1&ved=2ahUKEwja-Ka-zLSMAxXiHBAIHb1RGLAQM3oECBcQAA",
-    link: "https://google.com",
-    map_id: 1,
-    name: "Dust 2",
-    map_lineups: [
-        {
-            grenade_id: 1,
-            map_id: 1,
-            type_id: 1,
-            link_to_video: "https://youtu.be/example1",
-            user_id: 101,
-            created_at: "2025-03-31T15:10:14.890Z",
-            title: "Smoke CT Mirage",
-            description: "Смок на CT с T-спавна для выхода на A.",
-            is_approved: true,
-            views: 1500,
-            preview_image_link: "https://example.com/smoke_ct.jpg",
-            property_list: [
-                {
-                    key: "jump_throw",
-                    value: "true",
-                },
-            ],
-            grenade_class: {
-                name: "Smoke",
-                description: "Дымовая граната, блокирует видимость.",
-                price: 300,
-            },
-        },
-        {
-            grenade_id: 2,
-            map_id: 1,
-            type_id: 1,
-            link_to_video: "https://youtu.be/example2",
-            user_id: 102,
-            created_at: "2025-03-31T15:15:10.000Z",
-            title: "Jungle Smoke Mirage",
-            description: "Смок в Jungle с T-спавна.",
-            is_approved: true,
-            views: 1200,
-            preview_image_link: "https://example.com/smoke_jungle.jpg",
-            property_list: [
-                {
-                    key: "jump_throw",
-                    value: "true",
-                },
-            ],
-            grenade_class: {
-                name: "Smoke",
-                description: "Дымовая граната, блокирует видимость.",
-                price: 300,
-            },
-        },
-        {
-            grenade_id: 3,
-            map_id: 1,
-            type_id: 2,
-            link_to_video: "https://youtu.be/example3",
-            user_id: 103,
-            created_at: "2025-03-31T15:20:30.000Z",
-            title: "Molotov Default Mirage",
-            description: "Молотов на дефолтную позицию для выбивания плента A.",
-            is_approved: true,
-            views: 800,
-            preview_image_link: "https://example.com/molotov_default.jpg",
-            property_list: [],
-            grenade_class: {
-                name: "Molotov",
-                description: "Зажигательная граната, наносит урон по площади.",
-                price: 400,
-            },
-        },
-        {
-            grenade_id: 4,
-            map_id: 1,
-            type_id: 1,
-            link_to_video: "https://youtu.be/example4",
-            user_id: 104,
-            created_at: "2025-03-31T15:25:00.000Z",
-            title: "Long Doors Smoke Dust2",
-            description: "Смок на Лонг двери с T-спавна.",
-            is_approved: true,
-            views: 1000,
-            preview_image_link: "https://example.com/smoke_long_doors.jpg",
-            property_list: [],
-            grenade_class: {
-                name: "Smoke",
-                description: "Дымовая граната, блокирует видимость.",
-                price: 300,
-            },
-        },
-        {
-            grenade_id: 5,
-            map_id: 1,
-            type_id: 3,
-            link_to_video: "https://youtu.be/example5",
-            user_id: 105,
-            created_at: "2025-03-31T15:30:20.000Z",
-            title: "Flash Catwalk Dust2",
-            description: "Флешка с Т-спавна на Catwalk.",
-            is_approved: true,
-            views: 950,
-            preview_image_link: "https://example.com/flash_catwalk.jpg",
-            property_list: [
-                {
-                    key: "pop_flash",
-                    value: "true",
-                },
-            ],
-            grenade_class: {
-                name: "Flashbang",
-                description: "Светошумовая граната, ослепляет игроков.",
-                price: 200,
-            },
-        },
-        {
-            grenade_id: 6,
-            map_id: 1,
-            type_id: 1,
-            link_to_video: "https://youtu.be/example6",
-            user_id: 106,
-            created_at: "2025-03-31T15:35:50.000Z",
-            title: "B Main Smoke Inferno",
-            description: "Смок на B Main с CT-спавна.",
-            is_approved: true,
-            views: 1100,
-            preview_image_link: "https://example.com/smoke_b_main.jpg",
-            property_list: [
-                {
-                    key: "pop_flash",
-                    value: "true",
-                },
-            ],
-            grenade_class: {
-                name: "Smoke",
-                description: "Дымовая граната, блокирует видимость.",
-                price: 300,
-            },
-        },
-    ],
-}
