+++ tg-frontend/src/entities/map/index.ts
@@ -1,7 +1,7 @@
 export { api as mapApi } from "./api/client"
 export { testMapPageServer, testMapsServer } from "./api/__test-handler"
-export { mockMapPage, mockMaps } from "./api/__mocks"
+export { mockMapPage, mockMaps } from "./model/__mocks"
 export { MapCard } from "./ui/map-card/map-card"
 export { MapsList } from "./ui/maps-list"
-export type { MapModel, MapPageModel } from "./model"
-export { mapDTOschema, mapPageDTOschema } from "./model"
+export type { MapModel, MapPageModel } from "./model/domain"
+export { mapDTOschema, mapPageDTOschema } from "./model/domain"
