+++ tg-frontend/src/entities/map/model/domain.ts
