+++ tg-frontend/src/entities/map/ui/map-card/__mocks.ts
@@ -1,4 +1,4 @@
-import { MapModel } from "../../model"
+import { MapModel } from "../../model/domain"
 
 export const mapMock: MapModel = {
     imageLink:
